layout/style/test/test_cascade.html

<!DOCTYPE HTML>
<!-- vim: set shiftwidth=4 tabstop=8 autoindent expandtab: -->
<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is test_cascade.html
   -
   - The Initial Developer of the Original Code is the Mozilla Foundation.
   - Portions created by the Initial Developer are Copyright (C) 2008
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -   L. David Baron <dbaron@dbaron.org>, Mozilla Corporation (original author)
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the LGPL or the GPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->
<html>
<head>
  <title>Test for Author style sheet aspects of CSS cascading</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <style type="text/css">

  </style>
</head>
<body id="thebody">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<div class="content_class" id="content" style="position:relative"></div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Author style sheet aspects of CSS cascading **/

var style_element = document.createElement("style");
var style_contents = document.createTextNode("");
style_element.appendChild(style_contents);
document.getElementsByTagName("head")[0].appendChild(style_element);

var div = document.getElementById("content");
var cs = window.getComputedStyle(div, "");
var zindex = 0;

/**
 * Given the selectors |sel1| and |sel2|, in that order (the "order"
 * aspect of the cascade), with declarations that are !important if
 * |imp1|/|imp2| are true, assert that the one that wins in the
 * cascading order is given by |winning| (which must be either 1 or 2).
 */
function do_test(sel1, imp1, sel2, imp2, winning) {
  var ind1 = ++zindex;
  var ind2 = ++zindex;
  style_contents.data =
    sel1 + " { z-index: " + ind1 + (imp1 ? "!important" :"") + " } " +
    sel2 + " { z-index: " + ind2 + (imp2 ? "!important" :"") + " } ";
  var result = cs.zIndex;
  is(result, (winning == 1) ? ind1 : ind2,
     "cascading of " + style_contents.data);
}

// Test order, and order combined with !important
do_test("div", false, "div", false, 2);
do_test("div", false, "div", true, 2);
do_test("div", true, "div", false, 1);
do_test("div", true, "div", true, 2);

// Test specificity on a single element
do_test("div", false, "div.content_class", false, 2);
do_test("div.content_class", false, "div", false, 1);

// Test specificity across elements
do_test("body#thebody div", false, "body div.content_class", false, 1);
do_test("body div.content_class", false, "body#thebody div", false, 2);

// Test specificity combined with !important
do_test("div.content_class", false, "div", false, 1);
do_test("div.content_class", true, "div", false, 1);
do_test("div.content_class", false, "div", true, 2);
do_test("div.content_class", true, "div", true, 1);

function do_test_greater(sel1, sel2) {
  do_test(sel1, false, sel2, false, 1);
  do_test(sel2, false, sel1, false, 2);
}

function do_test_equal(sel1, sel2) {
  do_test(sel1, false, sel2, false, 2);
  do_test(sel2, false, sel1, false, 2);
}

// Test specificity of contents of :not()
do_test_equal("div.content_class", "div:not(.wrong_class)");
do_test_greater("div.content_class.content_class", "div.content_class");
do_test_greater("div.content_class", "div");
do_test_greater("div:not(.wrong_class)", "div");
do_test_greater("div:not(.wrong_class):not(.wrong_class)",
                "div:not(.wrong_class)");

</script>
</pre>
</body>
</html>
